Subsistemas 2 y 3/requirements.txt
argparse-addons==0.12.0
bidict==0.23.1
bitstruct==8.20.0
blinker==1.9.0
cantools==40.2.1
certifi==2025.4.26
click==8.1.8
crccheck==1.3.0
diskcache==5.6.3
dnspython==2.7.0
dominate==2.9.1
eventlet==0.39.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Executor==1.0.0
Flask-MQTT==1.2.1
Flask-SocketIO==5.5.1
greenlet==3.2.1
h11==0.16.0
influxdb-client==1.48.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==25.0
paho-mqtt==1.6.1
python-can==4.5.0
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
reactivex==4.0.4
setuptools==80.4.0
simple-websocket==1.1.0
six==1.17.0
textparser==0.24.0
typing_extensions==4.13.2
urllib3==2.4.0
visitor==0.1.3
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
